[ADD, ADJUST] Terminando de ajustar o bootstrap, e a conexão com o BD.

+++ View/alunos/alunosGrid.php
@@ -4,6 +4,7 @@ include '../../Config/conexao.php';
 include '../../public/header.php';
 include '../../public/body.php';
 
+
 $clientes = new Alunos($conexao);
 
 $dados = $clientes->fetchAll();
@@ -16,36 +17,44 @@ $dados = $clientes->fetchAll();
         <title></title>
     </head>
     <body>
-        <table class="table table-bordered table-hover" style="background-color: #c0c;">
-            <thead>
-                <tr>
-                    <th>#</th>
-                    <th>Nome</th>
-                    <th>CPF</th>
-                    <th>Estado</th>
-                    <th>Numero</th> 
-                    <th>Professores</th>
-                    <th>Editar</th>
-                    <th>Excluir</th>
-                </tr>
-            </thead>
-            <tbody>
-                <?php foreach ($dados as $key => $row) :
-                    ?>
+        <div class="col-md-12">
+            <table class="table table-bordered table-hover" style="background-color: #935D69;">
+                <thead>
                     <tr>
-                        <td><?= $row['id']; ?></td>
-                        <td><?= $row['nome']; ?></td>
-                        <td><?= $row['cpf']; ?></td>
-                        <td><?= $row['estado']; ?></td>
-                        <td><?= $row['numero']; ?></td>
-                        <td><?= $row['prof_id']; ?></td>
-                        <td><a href="../../View/alunos/alunosFormEditar.php?id=<?= $row['id'] ?>" class="btn btn-info">Editar</a></td>
-                        <td><a href="../../Controller/alunos/alunosDelete.php?id=<?= $row['id'] ?>" class="btn btn-danger">Excluir</a></td>
+                        <th>#</th>
+                        <th>Nome</th>
+                        <th>CPF</th>
+                        <th>Estado</th>
+                        <th>Numero</th> 
+                        <th>Professores</th>
+                        <?php if ($_SESSION['admin'] and $_SESSION['admin_2'] == 'true'): ?>
+                            <th>Editar</th>
+                            <th>Excluir</th>
+                        <?php endif; ?>
                     </tr>
-                <?php endforeach; ?>
-            </tbody>
-        </table>
-        <a href="../../View/alunos/alunosFormInserir.php" class="btn btn-success">Inserir</a>
+                </thead>
+                <tbody>
+                    <?php foreach ($dados as $key => $row) :
+                        ?>
+                        <tr>
+                            <td><?= $row['id']; ?></td>
+                            <td><?= $row['nome']; ?></td>
+                            <td><?= $row['cpf']; ?></td>
+                            <td><?= $row['estado']; ?></td>
+                            <td><?= $row['numero']; ?></td>
+                            <td><?= $row['nomeprof']; ?></td>
+                            <?php if ($_SESSION['admin'] and $_SESSION['admin_2'] == 'true'): ?>
+                                <td><a href="../../View/alunos/alunosFormEditar.php?id=<?= $row['id'] ?>" class="btn btn-info">Editar</a></td>
+                                <td><a href="../../Controller/alunos/alunosDelete.php?id=<?= $row['id'] ?>" class="btn btn-danger">Excluir</a></td>
+                            <?php endif; ?>
+                        </tr>
+                    <?php endforeach; ?>
+                </tbody>
+            </table>
+            <?php if ($_SESSION['admin'] or $_SESSION['admin_2'] == 'true'): ?>
+                <a href="../../View/alunos/alunosFormInserir.php" class="btn btn-success">Inserir</a>
+            <?php endif; ?>
+        </div>
     </body>
 </html>
